Have common operator map and allow for overrides
